Add shooty balls
As part of this, move stuff which was in Scene into BouncyScene, and add
a new CrimeScene which implements Scene.  Some of the ball-fondling code
moved from Scene into Ball.

There are also some changes I'm not sure about because they were made
several months ago; in particular, the pause business on
SceneUpdateThread looks like we busy-wait while we're paused.