issue #7: keep score (first pass)
Friday night Cafe Mokka update: keep score, display Toast when game
ends.  This is bad/temporary because we want a real Activity instead
of a short-lived Toast.

Also, we're not displaying the score when the game ends because a good
tile reached the bottom of the screen.  The bit which displays the
score at the end of the game should be moved into a new method, and
that method should be called when the person hits a bad tile (where
we're currently calling Toast.makeText()), and also when we decide the
game is over because a good tile hit the bottom of the screen.