issue #7: Add StartGameActivity & GameOverActivity
Sunday afternoon Mosgo's changes.  These new activities are the ones
described in issue #7; now, instead of launching TileActivity from
the MainActivity menu, we launch StartGameActivity; it's got a button
which starts the TileActivity; and, now, when the game is over,
TileActivity starts the GameOverActivity.  When that finishes,
TileActivity finishes too.

Also, as part of this, TileActivity starts off moving instead of
waiting for someone to tap on a "start" tile.

Note that the layouts of these two new activities are horrible & ugly
& temporary.