Preference menu cosmetic changes
Depending on whether timed tiles or normal tiles are selected, show/hide
preference menu items.  We couldn't figure out how to get this to work
with whole PreferenceCategory objects, so remove the PreferenceCategory
elements from the XML and (sadly) hard-code the list of which
preferences should be added/removed when the user toggles between timed
tiles & normal tiles.